// src/app/dynamic/model/FormType.ts
export enum FormType {
  CftpDeclaration = 1,
  RenewalDeclaration = 2,
  PlannedCpdEntry = 3,
  UnplannedCpdEntry = 4,
  PeerDiscussionEntry = 5,
  ReflectiveAccountEntry = 6,
  ExtenuatingCircumstances = 7,
  CftpDeclarationRos = 8,
  RenewalDeclarationRos = 9,
  CftpDeclarationPRos = 10,
  CftpDeclarationTros = 11,
  RenewalDeclarationTros = 12,
  CftpDeclarationARos = 13,
  RenewalDeclarationARos = 14,
  CftpDeclarationFinalTraining = 15,
  RenewalDeclarationFinalTraining = 16,
  CftpDeclarationVRForm = 17,
  CftpDeclarationRTR = 18,
  JoinDeclarationRTR = 19,
  CftpDeclarationCPSC = 20
}